fix(relationships): Fix relationship definitions
@see https://github.com/ubiquits/ubiquits/pull/123
