+++ package.json
@@ -16,7 +16,7 @@
     "@angular/router": "3.0.0-beta.2",
     "@angular/router-deprecated": "2.0.0-rc.2",
     "@angular/upgrade": "2.0.0-rc.4",
-    "@ubiquits/core": "^0.7.7",
+    "@ubiquits/core": "^0.7.13",
     "chalk": "^1.1.3",
     "dotenv": "^2.0.0",
     "lodash": "^4.12.0",
