+++ src/common/models/pet.model.ts
@@ -1,0 +1,28 @@
+import {
+  Model,
+  AbstractModel,
+  Primary,
+  UUID,
+  StoredProperty,
+  CreatedDate,
+  UpdatedDate,
+  HasOne,
+  MaxLength
+} from '@ubiquits/core/common';
+import { User } from './user.model';
+
+@Model({
+  storageKey: 'pets',
+})
+export class Pet extends AbstractModel {
+
+  @Primary()
+  public petId: string;
+
+  @StoredProperty()
+  public name: string;
+
+  @BelongsTo(f => User, user => user.userId)
+  public user:User;
+
+}
